requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask==3.0.0
gunicorn==21.2.0
argparse 